+++ extensions/cast/src/main/java/com/google/android/exoplayer2/ext/cast/CastPlayer.java
@@ -351,7 +351,8 @@ public final class CastPlayer extends BasePlayer {
     // We update the local state and send the message to the receiver app, which will cause the
     // operation to be perceived as synchronous by the user. When the operation reports a result,
     // the local state will be updated to reflect the state reported by the Cast SDK.
-    setPlayerStateAndNotifyIfChanged(playWhenReady, playbackState);
+    setPlayerStateAndNotifyIfChanged(
+        playWhenReady, PLAY_WHEN_READY_CHANGE_REASON_USER_REQUEST, playbackState);
     flushNotifications();
     PendingResult<MediaChannelResult> pendingResult =
         playWhenReady ? remoteMediaClient.play() : remoteMediaClient.pause();
@@ -625,8 +626,14 @@ public final class CastPlayer extends BasePlayer {
       newPlayWhenReadyValue = !remoteMediaClient.isPaused();
       playWhenReady.clearPendingResultCallback();
     }
+    @PlayWhenReadyChangeReason
+    int playWhenReadyChangeReason =
+        newPlayWhenReadyValue != playWhenReady.value
+            ? PLAY_WHEN_READY_CHANGE_REASON_REMOTE
+            : PLAY_WHEN_READY_CHANGE_REASON_USER_REQUEST;
     // We do not mask the playback state, so try setting it regardless of the playWhenReady masking.
-    setPlayerStateAndNotifyIfChanged(newPlayWhenReadyValue, fetchPlaybackState(remoteMediaClient));
+    setPlayerStateAndNotifyIfChanged(
+        newPlayWhenReadyValue, playWhenReadyChangeReason, fetchPlaybackState(remoteMediaClient));
   }
 
   @RequiresNonNull("remoteMediaClient")
@@ -718,13 +725,21 @@ public final class CastPlayer extends BasePlayer {
   }
 
   private void setPlayerStateAndNotifyIfChanged(
-      boolean playWhenReady, @Player.State int playbackState) {
-    if (this.playWhenReady.value != playWhenReady || this.playbackState != playbackState) {
-      this.playWhenReady.value = playWhenReady;
+      boolean playWhenReady,
+      @Player.PlayWhenReadyChangeReason int playWhenReadyChangeReason,
+      @Player.State int playbackState) {
+    boolean playWhenReadyChanged = this.playWhenReady.value != playWhenReady;
+    if (playWhenReadyChanged || this.playbackState != playbackState) {
       this.playbackState = playbackState;
+      this.playWhenReady.value = playWhenReady;
       notificationsBatch.add(
           new ListenerNotificationTask(
-              listener -> listener.onPlayerStateChanged(playWhenReady, playbackState)));
+              listener -> {
+                listener.onPlayerStateChanged(playWhenReady, playbackState);
+                if (playWhenReadyChanged) {
+                  listener.onPlayWhenReadyChanged(playWhenReady, playWhenReadyChangeReason);
+                }
+              }));
     }
   }